refactor(server): Fold server_config.h into server.h
