+++ include/open62541/plugin/historydatabase.h
@@ -8,7 +8,7 @@
 #ifndef UA_PLUGIN_HISTORYDATABASE_H_
 #define UA_PLUGIN_HISTORYDATABASE_H_
 
-#include <open62541/server.h>
+#include <open62541/util.h>
 
 _UA_BEGIN_DECLS
