+++ include/open62541/plugin/network.h
@@ -9,8 +9,8 @@
 #ifndef UA_PLUGIN_NETWORK_H_
 #define UA_PLUGIN_NETWORK_H_
 
+#include <open62541/util.h>
 #include <open62541/plugin/log.h>
-#include <open62541/server.h>
 
 _UA_BEGIN_DECLS
