+++ include/open62541/util.h
@@ -10,8 +10,23 @@
 
 #include <open62541/types.h>
 #include <open62541/types_generated.h>
+#include <open62541/types_generated_handling.h>
 
 _UA_BEGIN_DECLS
+
+/**
+ * Forward Declarations
+ * --------------------
+ * Opaque oointers used by the plugins. */
+
+struct UA_Server;
+typedef struct UA_Server UA_Server;
+
+struct UA_ServerConfig;
+typedef struct UA_ServerConfig UA_ServerConfig;
+
+struct UA_Client;
+typedef struct UA_Client UA_Client;
 
 /**
  * Endpoint URL Parser
